Extract eth_getLogs unit test from blockchain to filtermanager
